Factor out token verification data providers
