+++ tests/Functional/Token/Frontend/VerifyTest.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace App\Tests\Functional\Token\Frontend;
 
-use App\Tests\Functional\Token\AbstractTokenTest;
+use App\Tests\Functional\Token\AbstractVerifyTest;
 use App\Tests\Services\TestUserFactory;
 use Lexik\Bundle\JWTAuthenticationBundle\Encoder\JWTEncoderInterface;
 
-class VerifyTest extends AbstractTokenTest
+class VerifyTest extends AbstractVerifyTest
 {
     /**
      * @dataProvider verifyUnauthorizedDataProvider
@@ -18,24 +18,6 @@ class VerifyTest extends AbstractTokenTest
         $response = $this->application->makeFrontendVerifyTokenRequest($jwt);
 
         self::assertSame(401, $response->getStatusCode());
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function verifyUnauthorizedDataProvider(): array
-    {
-        return [
-            'no jwt' => [
-                'token' => null,
-            ],
-            'malformed jwt' => [
-                'token' => 'malformed.jwt.token',
-            ],
-            'invalid jwt' => [
-                'token' => 'eyJhbGciOiJIUzI1NiJ9.e30.ZRrHA1JJJW8opsbCGfG_HACGpVUMN_a9IV7pAx_Zmeo',
-            ],
-        ];
     }
 
     public function testVerifyValidJwt(): void
@@ -69,38 +51,9 @@ class VerifyTest extends AbstractTokenTest
         $encoder = self::getContainer()->get('lexik_jwt_authentication.encoder');
         \assert($encoder instanceof JWTEncoderInterface);
 
-        $verifyResponse = $this->application->makeApiVerifyTokenRequest($encoder->encode($tokenData));
+        $verifyResponse = $this->application->makeFrontendVerifyTokenRequest($encoder->encode($tokenData));
 
         self::assertSame($expectedResponseStatusCode, $verifyResponse->getStatusCode());
         self::assertStringContainsString($expectedResponseBodyContains, $verifyResponse->getBody()->getContents());
-    }
-
-    /**
-     * @return array<mixed>
-     */
-    public function verifyInvalidUserDataDataProvider(): array
-    {
-        return [
-            'empty token data' => [
-                'tokenData' => [],
-                'expectedResponseStatusCode' => 401,
-                'expectedResponseBodyContains' => 'Unable to find key \u0022email\u0022 in the token payload.',
-            ],
-            'payload sub key missing' => [
-                'tokenData' => [
-                    'email' => 'user@example.com',
-                ],
-                'expectedResponseStatusCode' => 500,
-                'expectedResponseBodyContains' => 'Payload key &quot;sub&quot; invalid',
-            ],
-            'payload roles key missing' => [
-                'tokenData' => [
-                    'email' => 'user@example.com',
-                    'sub' => 'user@example.com',
-                ],
-                'expectedResponseStatusCode' => 500,
-                'expectedResponseBodyContains' => 'Payload key &quot;roles&quot; invalid',
-            ],
-        ];
     }
 }
